+++ tests/doc/test_backward.py
@@ -13,6 +13,6 @@ def test_backward():
     y1 = torch.tensor([-1.0, 1.0], device=DEVICE) @ param
     y2 = (param**2).sum()
 
-    backward([y1, y2], [param], A=UPGrad())
+    backward([y1, y2], UPGrad())
 
     assert_close(param.grad, torch.tensor([0.5000, 2.5000], device=DEVICE), rtol=0.0, atol=1e-04)